+++ client/ui/HighlightsAnalyzed.jsx
@@ -18,7 +18,7 @@ class HighlightsAnalyzed extends Component {
       var that = this;
       setTimeout(function() {
           that.show();
-      }, 99);
+      }, 150);
   }
   show () {
       this.setState({hidden : ""});